+++ include/pinocchio/autodiff/casadi-algo.hpp
@@ -5,6 +5,7 @@
 #ifndef __pinocchio_autodiff_code_generator_algo_hpp__
 #define __pinocchio_autodiff_code_generator_algo_hpp__
 
+#include <casadi/core/casadi_types.hpp>
 #include <casadi/core/code_generator.hpp>
 
 #include "pinocchio/autodiff/casadi.hpp"
@@ -76,9 +77,12 @@ namespace pinocchio
         buildMap();
         // Generated code;
         cg_generated.add(ad_fun);
+        fun_operation_count = ad_fun.n_instructions() - ad_fun.nnz_in() - ad_fun.nnz_out();
         if (build_jacobian)
         {
           cg_generated.add(ad_fun_derivs);
+          fun_derivs_operation_count =
+            ad_fun_derivs.n_instructions() - ad_fun_derivs.nnz_in() - ad_fun_derivs.nnz_out();
         }
         cg_generated.generate();
       }
@@ -119,6 +123,16 @@ namespace pinocchio
         }
       }
 
+      casadi_int getFunOperationCount() const
+      {
+        return fun_operation_count;
+      }
+
+      casadi_int getFunDerivsOperationCount() const
+      {
+        return fun_derivs_operation_count;
+      }
+
     protected:
       ADModel ad_model;
       ADData ad_data;
@@ -133,6 +147,7 @@ namespace pinocchio
       ADFun ad_fun, ad_fun_derivs;
       ADFun fun, fun_derivs;
       std::vector<DMMatrix> fun_output, fun_output_derivs;
+      casadi_int fun_operation_count, fun_derivs_operation_count;
     };
 
     template<typename _Scalar>